fix: failing tests and rename services
